Rename - util 클래스 이름 및 패키지 구조 변경
