+++ src/main/java/com/trinity/ctc/domain/notification/entity/NotificationHistory.java
@@ -4,7 +4,7 @@ import com.google.firebase.messaging.MessagingErrorCode;
 import com.trinity.ctc.domain.notification.entity.result.SentResult;
 import com.trinity.ctc.domain.notification.entity.type.NotificationType;
 import com.trinity.ctc.domain.user.entity.User;
-import com.trinity.ctc.util.formatter.JsonConverter;
+import com.trinity.ctc.util.formatter.JsonUtil;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -28,7 +28,7 @@ public class NotificationHistory {
     @Enumerated(EnumType.STRING)
     private NotificationType type;
 
-    @Convert(converter = JsonConverter.class)
+    @Convert(converter = JsonUtil.class)
     private Map<String, String> message;
 
     private LocalDateTime sentAt;
